ensure Synced condition on Domain resources
Adds custom code to generic hook templates that places a
ConditionTypeResourceSynced on Domain resources with a status of False
when the Status.Processing field of the Domain is True and a status of
True otherwise.

Refactors the e2e tests to match the style of the RDS controller and
adds assertions for the ResourceSynced condition.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>