add checks for endpoint/endpoints field in status (#5)
The endpoint and endpoints field in the Domain.Status struct are filled
in when the domain's endpoint URL or URLs (if VPC endpoints) are
established. Adds an e2e test to verify these fields on the Status
struct are indeed populated, which after adding
requeue_on_success_seconds to the generator.yaml file, is indeed true.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.