bring controller up to modern runtime and deps (#49)
This controller has been languishing for quite some time with old ACK runtime and various dependencies (that dependabot keeps complaining about). This commit brings the controller up to the most modern runtime and set of upstream dependencies.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.