Regenerate controller using SDK v1.55.5 (#71)
Typical regeneration PR, bringing a lot of the recently introduced fields in OpenSearchService
controller. We noticed that there is a also a new introduced resource named `VPCEndpoint`,
that we will ignore for now.

There are some go tags generation issues in this PR, they will need to be fixed in
github.com/aws-controllers-k8s/pkg

/hold

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.