feat: set AutoTuneOptions during READ_ONE
Currently, the api we used to describe a domain resource did not return
the AutoTuneOption config, with all its values populated (maintenance
schedule, desiredState, etc.).
Without the latest state of AutoTuneOptions, updates were not
successful.

With this change, we will be adding an API call to retrieve the latest
AutoTuneOptions.